% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean.R
\name{as.boolean_list}
\alias{as.boolean_list}
\title{A method to generate boolean objects}
\usage{
as.boolean_list(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Additional options}
}
\description{
\code{as.boolean_list} returns a list of boolean objects.
}
\details{
This method converts objects into a list of the boolean terms
within the (model) object,
or an empty list if the supplied object has no corresponding
boolean terms.
}
