% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept.R
\name{as.intercept_list}
\alias{as.intercept_list}
\title{Return a list of intercept objects}
\usage{
as.intercept_list(x, ...)
}
\arguments{
\item{x}{Object to convert to intercepts}

\item{...}{Other options}
}
\value{
A list of intercept objects of length 0 or 1
}
\description{
\code{as.intercept_list} returns a list of intercept objects.
}
\details{
This method extracts the intercept information from a model
object.
If the object has an intercept,
the return value is a list of length one
whose inner object is an intercept object.
If the object does not have an intercept,
the return value is an empty list.
}
