% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{as.linear_list}
\alias{as.linear_list}
\title{A method to generate linear objects}
\usage{
as.linear_list(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Additional options}
}
\description{
\code{as.linear_list} returns a list of linear objects.
}
\details{
This method converts objects into a list of the linear terms
within the (model) object,
or an empty list if the supplied object has no corresponding
linear terms.
}
