% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{as.lookup_list}
\alias{as.lookup_list}
\title{A method to generate lookup objects}
\usage{
as.lookup_list(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Additional options}
}
\description{
\code{as.lookup_list} returns a list of lookup objects.
}
\details{
This method converts objects into a list of the lookup terms
within the (model) object,
or an empty list if the supplied object has no corresponding
lookup terms.
}
