% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean.R
\name{boolean}
\alias{boolean}
\title{Returns a boolean object}
\usage{
boolean(x, data = matrix())
}
\arguments{
\item{x}{Name of object}

\item{data}{matrix}
}
\description{
\code{boolean} returns an boolean object.
}
