% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{boruta_feature_selection}
\alias{boruta_feature_selection}
\title{Boruta Feature Selection}
\usage{
boruta_feature_selection(DT, CS, S)
}
\arguments{
\item{DT}{Data frame containing the features and outcome.}

\item{CS}{Censoring status.}

\item{S}{Survival time or binary indicator for survival.}
}
\value{
A data frame of selected features and their counts.
}
\description{
Applies the Boruta feature selection algorithm to the dataset.
}
