% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_plot.R
\name{cal_plot}
\alias{cal_plot}
\title{Visualise calibration}
\usage{
cal_plot(
  predictr = NULL,
  fit = NULL,
  risk_ntile = NULL,
  risk_class = NULL,
  se = F
)
}
\arguments{
\item{predictr}{Output from the predictr function (alternative to fit parameter)}

\item{fit}{Logistic regression fit object (alternative to predictr parameter)}

\item{risk_ntile}{Numerical value specifying the number of predictive risk quantiles (default=NULL)}

\item{risk_class}{Numerical list specifying the cutoffs for predictive risk classes (default=NULL; range: 0 to 1)}

\item{se}{Logical value specifying if confidence intervals should be added to the plot}
}
\value{
ggplot of the calibration plot with predicted risk either continuous (default) or categoried by quantiles or risk classes.
}
\description{
Use to obtain a calibration plot for a predictive model
}
