% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_table.R
\name{cal_table}
\alias{cal_table}
\title{Table of calibration}
\usage{
cal_table(predictr = NULL, fit = NULL, risk_ntile = 10, risk_class = NULL)
}
\arguments{
\item{predictr}{Output from the predictr function (alternative to fit parameter)}

\item{fit}{Logistic regression fit object (alternative to predictr parameter)}

\item{risk_ntile}{Numerical value specifying the number of predictive risk quantiles (default=10)}

\item{risk_class}{Numerical list specifying the cutoffs for predictive risk classes (default=NULL; range: 0 to 1)}
}
\value{
Tibble of the number of events / sample per risk quantile or risk classes
}
\description{
Use to obtain a calibration table for a predictive model
}
