% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_features_to_observations_ratio}
\alias{check_features_to_observations_ratio}
\title{Check Features to Observations Ratio}
\usage{
check_features_to_observations_ratio(df, N = 10)
}
\arguments{
\item{df}{A dataframe to be checked.}

\item{N}{A positive numeric value representing the threshold ratio (1/N).}
}
\value{
TRUE if the number of features is less than 1/N times the number of observations, FALSE otherwise.
}
\description{
This function checks if the number of features in a dataframe is less than 1/N times the number of observations, where N is a user-defined parameter.
}
