% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Check for Missing Values in a CSV File}
\usage{
check_missing_values(file_path)
}
\arguments{
\item{file_path}{The file path to the CSV file to be checked.}
}
\value{
TRUE if the CSV file contains missing values (NAs), FALSE otherwise.
}
\description{
This function reads a CSV file from the specified file path, checks if it contains any missing values (NAs), and returns TRUE if any missing values are found.
}
