% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient.R
\name{coefficient}
\alias{coefficient}
\title{Extract coefficents from a logistic regression object}
\usage{
coefficient(fit, coefficient = "beta")
}
\arguments{
\item{fit}{Logistic regression fit object}

\item{coefficient}{String of column name in df which represents the end date.}
}
\value{
Tibble of labelled coefficents
}
\description{
Use to extract coefficents from a logistic regression object
}
