% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-price.R
\name{contract_interval}
\alias{contract_interval}
\title{Convert contract names to factor intervals}
\usage{
contract_interval(x)
}
\arguments{
\item{x}{A character vector of interval contract names.}
}
\value{
A factor vector.
}
\description{
Convert contract names to factor intervals
}
\examples{
contract_interval(c("1\%-", "4\%+", "2\% - 3\%", "1\% - 2\%", "3\% - 4\%"))
contract_interval(c("219-", "220 - 229", "240+",  "230 - 239"))
}
