% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_correlation_heatmap}
\alias{create_correlation_heatmap}
\title{Create Correlation Heatmap}
\usage{
create_correlation_heatmap(DT, plotName, method = "pearson")
}
\arguments{
\item{DT}{Data frame to visualize correlations.}

\item{plotName}{Name of the output plot file.}

\item{method}{Correlation method to use.}
}
\description{
Generates a heatmap for visualizing correlations in the data.
}
