% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eval_xgb_rmpse.R
\name{eval_xgb_rmpse}
\alias{eval_xgb_rmpse}
\title{Custom Evaluation Functions for xgboost}
\usage{
eval_xgb_rmpse(preds, dtrain)
}
\arguments{
\item{preds}{predictions}

\item{dtrain}{DMatrix of the reference data}
}
\description{
This function loads various xgboost evaluation function for xgboost
}
\examples{
eval_xgb_rmpse(preds, dtrain)
eval_xgb_kappa(preds, dtrain)
}

