% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_prediction}
\alias{example_prediction}
\title{Example dataset of model predictions}
\format{
A data frame with 600 rows and 3 variables.
}
\usage{
data(example_prediction)
}
\description{
Example dataset of model predictions for use within vignettes
}
\keyword{data}
