% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_probast.R
\docType{data}
\name{example_probast}
\alias{example_probast}
\title{Example dataset of PROBAST assessment}
\format{
A data frame with 36 rows and 25 variables.
}
\usage{
data(example_probast)
}
\description{
Example dataset of PROBAST assessment for use within vignettes
}
\keyword{data}
