% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_tripod.R
\docType{data}
\name{example_tripod}
\alias{example_tripod}
\title{Example dataset of TRIPOD assessment}
\format{
A data frame with 36 rows and 39 variables.
}
\usage{
data(example_tripod)
}
\description{
Example dataset of TRIPOD assessment for use within vignettes
}
\keyword{data}
