% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_modality_from_filename}
\alias{extract_modality_from_filename}
\title{extract_modality_from_filename}
\usage{
extract_modality_from_filename(filename)
}
\arguments{
\item{filename}{A string representing the full path of the file, including its name and extension.}
}
\value{
A string representing the modality extracted from the filename. For example, if the input
filename is "path/to/feature_modality.csv", the function will return "modality".
}
\description{
This function is designed to extract the modality from a filename. It assumes that the filename
format includes the modality as the second last part, followed by a file extension (e.g., .csv).
}
\examples{
filename <- "/path/to/your/file/feature_modality.csv"
modality <- group_methods(filename)
print(modality)  # Should print 'modality'

}
