% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{extract_significant_features}
\alias{extract_significant_features}
\title{Extract Significant Features from PCA}
\usage{
extract_significant_features(
  pca_result,
  num_components,
  threshold,
  ignore_column = "sub"
)
}
\arguments{
\item{pca_result}{The result of PCA computation using prcomp or similar methods.}

\item{num_components}{The number of principal components to consider for feature extraction.}

\item{threshold}{The threshold for determining significance of features based on their loadings.}

\item{ignore_column}{Optional; the name of the column to be ignored during feature extraction, defaults to "sub".}
}
\value{
A vector of significant feature names.
}
\description{
This function extracts significant features from the PCA result based on the specified number of components and threshold. It can optionally ignore a specific column during the extraction process.
}
\examples{
pca_res <- prcomp(DT, scale. = TRUE)
significant_features <- extract_significant_features(pca_res, num_components = 5, threshold = 0.3)
}
