% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{feature_selection_loop}
\alias{feature_selection_loop}
\title{Feature Selection Loop}
\usage{
feature_selection_loop(DT, outputDir, nrep = 1000, run_parallel = TRUE)
}
\arguments{
\item{DT}{Data frame containing the features and outcome variables.}

\item{outputDir}{Directory to save the output files.}
}
\value{
The function does not return a value but saves feature selection results to files in the specified output directory.
}
\description{
This function executes a feature selection loop for a given dataset. It applies univariate feature selection, Boruta feature selection, and Lasso feature selection iteratively for a specified number of iterations.
}
