% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_correlated_features}
\alias{filter_correlated_features}
\title{Filter Highly Correlated Features}
\usage{
filter_correlated_features(df, cutoff)
}
\arguments{
\item{cutoff}{Correlation cutoff value to determine feature removal.}

\item{DT}{Data frame to perform correlation filtering on.}
}
\value{
Data frame with highly correlated features removed.
}
\description{
Removes features that are highly correlated based on the provided cutoff.
}
