% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{find_elbow_point}
\alias{find_elbow_point}
\title{Find the elbow point in a vector of values}
\usage{
find_elbow_point(values)
}
\arguments{
\item{values}{A numeric vector of values.}
}
\value{
An integer indicating the index of the elbow point.
}
\description{
This function calculates the elbow point in a vector of numeric values.
}
\examples{
data_values <- c(10, 15, 20, 25, 30, 35, 40, 45, 50)
find_elbow_point(data_values)

}
