% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{find_flattening_point}
\alias{find_flattening_point}
\title{Find the flattening point in a vector of variance values}
\usage{
find_flattening_point(variance, threshold = 0.01)
}
\arguments{
\item{variance}{A numeric vector of variance values.}

\item{threshold}{A numeric threshold defining the minimum incremental increase
in variance to be considered as a non-flattening point (default: 0.01).}
}
\value{
An integer indicating the flattening point index.
}
\description{
This function calculates the flattening point in a vector of variance values
based on the specified threshold.
}
\examples{
variance_values <- c(0.1, 0.2, 0.3, 0.31, 0.32, 0.33, 0.35, 0.4)
find_flattening_point(variance_values, threshold = 0.05)

}
