% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsImpute.R
\name{fsImpute}
\alias{fsImpute}
\title{Impute Missing Data}
\usage{
fsImpute(outputDir)
}
\arguments{
\item{outputDir}{directory where the significant feature results is stored from which a list of dataframes with missing values to be imputed.}
}
\description{
This function takes a list of dataframes and performs multiple imputation methods
on each dataframe. It returns a list of dataframes where each dataframe is imputed
using a different method.
}
\examples{
# Assuming df1 and df2 are dataframes with missing values
my_dfs <- list(df1, df2)
imputed_dfs <- impute_data(my_dfs)
names(imputed_dfs) # Lists the names of imputed dataframes
}
