% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsSurv.R
\name{fsSurv}
\alias{fsSurv}
\title{Main Survival Analysis Function}
\usage{
fsSurv(datacsv, outcomecsv, outcometype, outputDir_md, nrep, run_parallel)
}
\arguments{
\item{datacsv}{Path to the CSV file containing the data.}

\item{outcomecsv}{Path to the CSV file containing outcome data.}

\item{outcometype}{The type of outcome to analyze ('os', 'pfs', 'tox', or 'group').}

\item{outputDir}{Directory to save the output files.}
}
\description{
This function performs survival analysis based on the provided dataset and outcome type.
It integrates various feature selection methods like univariate analysis, Boruta, and Lasso.
}
