% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_models.R
\name{generate_models}
\alias{generate_models}
\title{Feature Selection Signature Analysis}
\usage{
generate_models(outputDir)
}
\arguments{
\item{outputDir}{Directory where the output dataframes will be saved.}
}
\value{
A list of dataframes processed based on the provided feature set and criteria.
}
\description{
This function performs feature selection signature analysis on a set of data files. It loads data, performs feature filtering based on frequency, and combines all features to create a comprehensive dataframe. The function then saves the processed dataframes into the specified output directories.
}
