% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{get_featuresUnivCox}
\alias{get_featuresUnivCox}
\title{Univariate Feature Selection for Cox Proportional Hazards Model}
\usage{
get_featuresUnivCox(DT, surv, p, NF)
}
\arguments{
\item{DT}{Data frame containing the features and surv object.}

\item{surv}{Survival object (created using Surv function).}

\item{p}{The p-value for significance testing.}

\item{NF}{Number of features to select.}
}
\value{
A list containing the selected features.
}
\description{
Performs univariate feature selection using Cox proportional hazards model.
}
