% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{get_featuresUnivGlm}
\alias{get_featuresUnivGlm}
\title{Univariate Feature Selection for Generalized Linear Models}
\usage{
get_featuresUnivGlm(DT, group, p, NF)
}
\arguments{
\item{DT}{Data frame containing the features and group.}

\item{group}{The group variable for analysis.}

\item{p}{The p-value for significance testing.}

\item{NF}{Number of features to select.}
}
\value{
A list containing the selected features.
}
\description{
Performs univariate feature selection using generalized linear models.
}
