% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_list}
\alias{get_list}
\title{Retrieve a Subset of CSV Lists Based on Included Names}
\usage{
get_list(datacsvlists, to_include)
}
\arguments{
\item{datacsvlists}{A list of CSV file paths.}

\item{to_include}{A vector of names to be included in the returned list.}
}
\value{
A filtered list of CSV file paths.
}
\description{
This function filters a list of CSV file paths, returning only those
whose associated names are included in a specified list.
}
