% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_mdroiname}
\alias{get_mdroiname}
\title{Extract name information from a CSV file path}
\usage{
get_mdroiname(datacsv)
}
\arguments{
\item{datacsv}{A character string representing the CSV file path.}
}
\value{
A character string containing the extracted name information.
}
\description{
This function extracts name information from a CSV file path by splitting
the filename based on underscores and returning a concatenated string.
}
