% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{getIsoLR}
\alias{getIsoLR}
\title{Get couples of (FPF, TPF) with the same LR}
\usage{
getIsoLR(FPF, TPF)
}
\arguments{
\item{FPF}{False Positive Fraction}

\item{TPF}{True Positive Fraction}
}
\value{
a tibble with intercept and slope for each LR
}
\description{
Get couples of (FPF, TPF) with the same LR
}
