% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{getIsoOR}
\alias{getIsoOR}
\title{Get couples of (FPF, TPF) with the same OR}
\usage{
getIsoOR(FPF, TPF)
}
\arguments{
\item{FPF}{False Positive Fraction}

\item{TPF}{True Positive Fraction}
}
\value{
a tibble with of FPF and TPF values
}
\description{
Get couples of (FPF, TPF) with the same OR
}
