% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{getOR}
\alias{getOR}
\title{Get the odds ratio from (FPF, TPF)}
\usage{
getOR(FPF, TPF)
}
\arguments{
\item{FPF}{False Positive Fraction}

\item{TPF}{True Positive Fraction}
}
\value{
a value
}
\description{
Get the odds ratio from (FPF, TPF)
}
\examples{
getOR(.4, .6)
}
