% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{lasso_feature_selection}
\alias{lasso_feature_selection}
\title{Lasso Feature Selection}
\usage{
lasso_feature_selection(DT, CS, S, cv_iter = 200, alpha = 1)
}
\arguments{
\item{DT}{Data frame containing the features and outcome.}

\item{CS}{Censoring status.}

\item{S}{Survival time or binary indicator for survival.}
}
\value{
A data frame of selected features and their counts.
}
\description{
Performs Lasso feature selection on the given dataset.
}
