% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load and Preprocess Data}
\usage{
load_data(datacsv, scaledf = FALSE, corrm = TRUE, appendname = FALSE)
}
\arguments{
\item{datacsv}{Path to the CSV file containing the data.}

\item{scaledf}{Boolean indicating whether to scale the data.}

\item{corrm}{Boolean indicating whether to apply a correlation matrix.}

\item{appendname}{Boolean indicating whether to append names to the data.}

\item{rmvna}{Boolean indicating whether to remove columns with NA or infinite values.}
}
\value{
A preprocessed data frame.
}
\description{
Loads and preprocesses the data from a given CSV file.
}
