% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup}
\alias{lookup}
\title{Returns a lookup object}
\usage{
lookup(x, data, levels)
}
\arguments{
\item{x}{Name of input variable}

\item{data}{Lookup table values}

\item{levels}{Levels in lookup table}
}
\description{
\code{lookup} returns a lookup object.
}
