% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-history.R
\name{market_history}
\alias{market_history}
\title{Market price history}
\format{
A tibble with 10 variables:
\describe{
\item{time}{The hour or day of price}
\item{mid}{Market ID}
\item{market}{Market question}
\item{cid}{Contract ID}
\item{contract}{Question answers}
\item{open}{Price at start of hour/day}
\item{close}{Price at end of hour/day}
\item{low}{Low price of hour/day}
\item{high}{High price of hour/day}
\item{volume}{Number of shares traded per hour/day}
}}
\usage{
market_history(mid, hourly = FALSE)
}
\arguments{
\item{mid}{The integer market ID.}

\item{hourly}{24 hourly rows per contract be returned? If not, 90 daily rows.}
}
\value{
A data frame of market contract prices over time.
}
\description{
Produce a data frame of contract prices over time, either hourly or daily.
For hourly tables, there is one row per contract per hour for the last 24
hours. For daily tables, one row per contract per day.
}
\examples{
market_history(mid = 3633)
market_history(mid = 6400)
}
