% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-plot.R
\name{market_plot}
\alias{market_plot}
\title{Plot market types}
\usage{
market_plot(mid, type = c("line", "bar"), n = 4)
}
\arguments{
\item{mid}{The integer market ID.}

\item{type}{One of "line" or "bar", calls either \code{\link[=market_history]{market_history()}} or
\code{\link[=market_price]{market_price()}}. For price history plots, \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} is used;
for current prices, \code{\link[ggplot2:geom_col]{ggplot2::geom_col()}} is used.}

\item{n}{If "line", the number of current top contracts lines highlight.}
}
\value{
A ggplot object.
}
\description{
Plot market types
}
\examples{
market_plot(mid = 3633, type = "line")
market_plot(mid = 6405, type = "bar")
}
