% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-price.R
\name{market_price}
\alias{market_price}
\title{Single market prices}
\usage{
market_price(mid)
}
\arguments{
\item{mid}{The integer market ID.}
}
\value{
A data frame of market contracts prices.
}
\description{
Single market prices
}
\examples{
market_price(mid = 3633)
market_price(mid = 6400)
market_price(mid = 6360)
}
