% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{merge_grouped_files}
\alias{merge_grouped_files}
\title{Merge Grouped Files}
\usage{
merge_grouped_files(file_group)
}
\arguments{
\item{file_group}{A character vector containing file paths to be merged.}
}
\value{
A merged dataframe containing data from all the input files.
}
\description{
This function takes a vector of file paths, reads the CSV files, and merges them by a common column ('sub').
}
