% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{mrmr_feature_selection}
\alias{mrmr_feature_selection}
\title{Minimum Redundancy Maximum Relevance Feature Selection}
\usage{
mrmr_feature_selection(DT, CS, S, N = 10, solution_count = 100)
}
\arguments{
\item{DT}{Data frame containing the features and the response variable.}

\item{response_col}{Name of the column in DT that contains the response variable.}
}
\value{
A vector of selected feature names based on the mRMR criteria.
}
\description{
Performs feature selection using the Minimum Redundancy Maximum Relevance (mRMR) method.
mRMR selects features that are highly relevant to the target response while also being minimally redundant with each other. This method is particularly useful in high-dimensional data scenarios, such as genomics and radiomics.
}
\examples{
# Assuming DT is your data frame and 'response' is the response variable
selected_features <- mrmr_feature_selection(DT, "response")
}
