% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi2ova.R
\name{multi2ova}
\alias{multi2ova}
\title{Convert multi-class targets into One-vs-All targets}
\usage{
multi2ova(y)
}
\arguments{
\item{y}{target (multi-class factor)}
}
\description{
This function determines the best combination of predictors.
}
\examples{
## T.B.A.

}

