% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{perform_pca_feature_selection}
\alias{perform_pca_feature_selection}
\title{Perform PCA Feature Selection}
\usage{
perform_pca_feature_selection(DT, apply_varimax = FALSE)
}
\arguments{
\item{DT}{Data frame containing the features for PCA.}
}
\value{
A list of significant features for each method and threshold.
}
\description{
This function performs PCA on the given dataset and extracts significant features using various methods including Kaiser's rule, the elbow method, and flattening point method, at different significance thresholds.
}
\examples{
pca_features <- perform_pca_feature_selection(DT)
}
