% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predictor}
\alias{predictor}
\title{Returns a predictor object}
\usage{
predictor(x, ...)
}
\arguments{
\item{x}{(Model) object}

\item{...}{Additional options}
}
\value{
A predictor object corresponding with \code{x}
}
\description{
\code{predictor} returns a predictor object
}
