% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictr.R
\docType{package}
\name{predictr}
\alias{predictr}
\alias{predictr-package}
\title{predictr: A package for fast model predictions}
\description{
The predictr package provides adaptors which convert standard
model objects into a new type of object.
This new object stores the minimal amount of data required to
run model predictions,
and these new predictions reproduce the results of the
\code{predict} method with greater speed and with less memory
usage.
}
