% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{prettyProp}
\alias{prettyProp}
\title{Prettify proportions}
\usage{
prettyProp(prop, digits = 0)
}
\arguments{
\item{prop}{A proportion}

\item{digits}{passed to round()}
}
\value{
A pretty string
}
\description{
Prettify proportions
}
\examples{
prettyProp(.501)
}
