% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{prettyRatio}
\alias{prettyRatio}
\title{Prettify ratios}
\usage{
prettyRatio(ratio, digits = 2)
}
\arguments{
\item{ratio}{A ratio}

\item{digits}{passed to signif()}
}
\value{
A pretty string
}
\description{
Prettify ratios
}
\examples{
prettyRatio(1/2.41)
}
