% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probast_format.R
\name{probast_format}
\alias{probast_format}
\title{Format PROBAST data prior to plotting}
\usage{
probast_format(data, id = "id", type = "type")
}
\arguments{
\item{data}{PROBAST data in format as specified in  "example_probast.csv".}

\item{id}{String specifying the column name containing the identifier for the specific paper/model.}

\item{type}{String specifying the column name containing the type of model ("d"=derivation, "dv" = "derivation and validation", "v" = validation)}
}
\value{
Tibble of PROBAST data formatted for a PROBAST plot / table
}
\description{
Format PROBAST data prior to plotting
}
