% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probast_plot.R
\name{probast_plot}
\alias{probast_plot}
\title{Plot PROBAST data}
\usage{
probast_plot(data)
}
\arguments{
\item{data}{Output from probast_table.}
}
\value{
PROBAST plot.
}
\description{
Plot PROBAST data
}
