% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probast_table.R
\name{probast_table}
\alias{probast_table}
\title{Format PROBAST risk of bias data as a table}
\usage{
probast_table(data)
}
\arguments{
\item{data}{Output from probast_format.}
}
\value{
Tibble of PROBAST risk of bias data
}
\description{
Format PROBAST risk of bias data as a table
}
