% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quick_tune.R
\name{quick_tune}
\alias{quick_tune}
\title{Quick paramemters tuning using caret}
\usage{
quick_tune(x, y, method = c("rf"), n_samp = 1000, n_cv_fold = 3,
  n_cv_repeat = 3, n_run = 25, n_tuneLength = 5, custom_tuneGrid = NULL,
  n_core = 7)
}
\arguments{
\item{x}{predictors}

\item{y}{targets}
}
\description{
Quick paramemters tuning using caret
}
\examples{
## T.B.A.
}

