% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_duplicate_signfeat}
\alias{remove_duplicate_signfeat}
\title{Filter and Rename DataFrame Columns Based on Maximum Count}
\usage{
remove_duplicate_signfeat(df)
}
\arguments{
\item{df}{A dataframe with column names that include a base name and a numeric suffix.}
}
\value{
A dataframe with filtered and renamed columns.
}
\description{
This function processes a dataframe by identifying columns with similar base names and different numeric suffixes. It retains only the column with the highest numeric suffix for each base name and renames it to include this maximum count.
}
\examples{
df <- data.frame(
  "feature1_123" = 1:10,
  "feature1_456" = 1:10,
  "feature2_789" = 21:30
)
filtered_df <- filter_columns(df)
}
