% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_highcorr.R
\name{remove_highcorr}
\alias{remove_highcorr}
\title{Remove highly correlated predictors}
\usage{
remove_highcorr(x, cutoff = 0.75)
}
\arguments{
\item{x}{a data frame or matrix of predictors (x).}

\item{cutoff}{cutoff correlation value (ideally >0.75)}
}
\description{
This function finds and removes attributes that are highly corrected.
}
\examples{
## T.B.A.

}

