% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{rf_feature_selection}
\alias{rf_feature_selection}
\title{Random Forest Feature Selection}
\usage{
rf_feature_selection(DT, CS, S)
}
\arguments{
\item{DT}{A data frame containing the features along with survival time (`S`) and censoring status (`CS`).}

\item{CS}{A vector indicating the censoring status. If your data is survival data, this parameter should be provided. For binary classification, it can be ignored.}

\item{S}{A vector indicating the survival time or the binary outcome. Survival time is expected if `CS` is provided, otherwise a binary outcome is assumed.}
}
\value{
A vector of selected feature names ranked by their importance.
}
\description{
This function implements the Random Forest Importance method for feature selection. It utilizes the Random Forest algorithm to assess the importance of each feature in the dataset. The function is designed to handle both binary and survival outcome data.
}
\examples{
# For binary classification
selected_features_binary <- random_forest_feature_selection(DT = your_data, S = your_binary_outcome)

# For survival data
selected_features_survival <- random_forest_feature_selection(DT = your_data, CS = your_censoring_status, S = your_survival_time)

}
