% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_metric.R
\name{roc_metric}
\alias{roc_metric}
\title{Derive diagnostic/prognostic accuracy metrics for one or more models.}
\usage{
roc_metric(data, model = NULL, event = "event", predict = "prediction")
}
\arguments{
\item{data}{Dataframe with at least 2 columns: (1) binary outcome of interest, (2) predicted probability of outcome (binary or continuous)}

\item{model}{Optional string specifying the column name specifying which model each outcome/prediction belongs to (default ="model"). This allows multiple models to be assessed within 1 function.}

\item{event}{String specifying the column name containing the binary outcome of interest (default ="event")}

\item{predict}{String specifying the column name containing predicted probability of outcome (default ="prediction")}
}
\value{
Tibble with relevant diagnostic/prognostic accuracy metrics for one or more models.
}
\description{
Derive diagnostic/prognostic accuracy metrics for one or more models.
}
