% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc_plot}
\alias{roc_plot}
\title{Plot ROC plot for one or more models.}
\usage{
roc_plot(data, confint = F)
}
\arguments{
\item{data}{Output from roc_plot_format}

\item{confint}{Logical value specifying if the confidence interval of the sensitivity/specificity should be displayed (must be same as in roc_plot_format)}
}
\value{
ggplot ROC plot with one or more models displayed
}
\description{
Plot ROC plot for one or more models.
}
