% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot_format.R
\name{roc_plot_format}
\alias{roc_plot_format}
\title{Format data prior to plotting ROC plot for one or more models.}
\usage{
roc_plot_format(
  data,
  model = NULL,
  event = "event",
  predict = "predict",
  smooth = F,
  confint = F
)
}
\arguments{
\item{data}{Dataframe with at least 2 columns: (1) binary outcome of interest, (2) predicted probability of outcome (binary or continuous)}

\item{model}{Optional string specifying the column name specifying which model each outcome/prediction belongs to (default ="model"). This allows multiple models to be assessed within 1 function.}

\item{event}{String specifying the column name containing the binary outcome of interest (default ="event")}

\item{predict}{String specifying the column name containing predicted probability of outcome (default ="prediction")}

\item{smooth}{Logical value specifying if the ROC curve should be smoothed (default=FALSE)}

\item{confint}{Logical value specifying if the confidence interval of the sensitivity/specificity should be calculated (default = FALSE)}
}
\value{
Tibble with sensitivity/specificity for one or more models.
}
\description{
Format data prior to plotting ROC plots for one or more models.
}
