% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save_dataframes}
\alias{save_dataframes}
\title{Save Dataframes to CSV Files}
\usage{
save_dataframes(dfs, output_dir, name = NULL)
}
\arguments{
\item{dfs}{A list of dataframes, where each dataframe is associated with a method.}

\item{output_dir}{A string representing the directory path where the CSV files will be saved. The function will create the directory if it does not exist.}

\item{name}{An optional string used as part of the filename for each saved dataframe. If provided, it is concatenated with the method name to generate the full filename. If not provided, the filename will consist only of the method name.}
}
\value{
No return value; this function is called for its side effect of saving files.
}
\description{
This function iterates over a list of dataframes, each associated with a method, and saves them as CSV files in a specified output directory. The filenames are generated using a template 'sigfeats_method_name.csv', where 'sigfeats' is an acronym for 'significant features', and 'method' is replaced with the actual method name.
}
