% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save_feature_files}
\alias{save_feature_files}
\title{Save Feature Files}
\usage{
save_feature_files(feature_data, outputDir)
}
\arguments{
\item{feature_data}{A named list of data frames or lists of data frames.}

\item{outputDir}{Directory to save the output files.}
}
\description{
Saves feature data to CSV files with names derived from the provided names.
}
