% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_fea_xgb.R
\name{select_fea_xgb}
\alias{select_fea_xgb}
\title{Feature selection using xgboost}
\usage{
select_fea_xgb(x, y, n_p = 1, t_obj = "multi:softprob",
  t_eval = "mlogloss", n_class = NULL, n_depth = 9, n_eta = 0.1,
  n_subsamp = 0.9, n_core = 8, n_round = 100, n_cv_fold = 3,
  n_final_cv_fold = 10, n_size = "auto", n_repeat = 3, verbose = TRUE)
}
\arguments{
\item{x}{predictors}

\item{y}{target}
}
\description{
Feature selection using xgboost
}
\examples{
## T.B.A.

}

