% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SL_train_xgb.R
\name{SL_train_reg_xgb}
\alias{SL_train_reg_xgb}
\title{Super Learner Training - xgboost}
\usage{
SL_train_reg_xgb(...)
}
\arguments{
\item{x}{predictors}

\item{ratio}{split ratio i.e. train:tune:hold = 80:10:10}
}
\description{
This function splits data into training/tuning/holdation
}
\examples{
a <- SL_train_reg_xgb(x, y)
}

