% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_x.R
\name{split_by_x}
\alias{split_by_x}
\title{Spliting data into training/tuning/holdation based on predictors (x)}
\usage{
split_by_x(x, ratio = c(8, 1, 1))
}
\arguments{
\item{x}{predictors}

\item{ratio}{split ratio i.e. train:tune:hold = 80:10:10}
}
\description{
This function splits data into training/tuning/holdation
}
\examples{
row_samp <- split_by_x(x)

}

