% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fuData.R
\name{transform_fuData}
\alias{transform_fuData}
\title{Transform longitudinal data into time-to-event format}
\usage{
transform_fuData(
  file_name,
  start_date,
  fu_date,
  var,
  id_loc = 1,
  units = "days",
  date_format = "\%m/\%d/\%Y"
)
}
\arguments{
\item{file_name}{The name of the file containing the longitudinal data.}

\item{start_date}{The name of the variable containing the start date of the observation period.}

\item{fu_date}{The prefix used for the variables containing the follow-up dates.}

\item{var}{The prefix used for the variables containing the variable of interest.}

\item{id_loc}{The column index or name containing the unique identifier for each observation (default is 1).}

\item{units}{The time units used for calculating the time to event (default is "days").}

\item{date_format}{The date format used in the input data (default is "%m/%d/%Y").}
}
\value{
A data frame in time-to-event format with one row per timepoint.
}
\description{
Transform longitudinal data into time-to-event format
}
