% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_features_df}
\alias{update_features_df}
\title{Update Feature Counts in a Data Frame}
\usage{
update_features_df(feature_df, new_features)
}
\arguments{
\item{feature_df}{A data frame with columns 'f' for features and 'count' for their counts.}

\item{new_features}{A vector of new features to be added or updated in the feature_df.}
}
\value{
The updated data frame with counts of features.
}
\description{
Updates the count of each feature in a data frame. If a feature is new, it is added to the data frame
with a count of 1. If a feature already exists, its count is incremented.
}
\examples{
feature_df <- data.frame(f = character(), count = integer())
new_features <- c("feature1", "feature2")
updated_df <- update_features_df(feature_df, new_features)
}
