% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_features_df_list}
\alias{update_features_df_list}
\title{Update Feature Counts in a List of Data Frames}
\usage{
update_features_df_list(pca_feature_counts, pca_features)
}
\arguments{
\item{pca_feature_counts}{A list of data frames, each representing feature counts for a specific PCA method.}

\item{pca_features}{A list of new features corresponding to each PCA method.}
}
\value{
A list of data frames with updated feature counts.
}
\description{
This function iterates over a list of feature data frames and a corresponding list of new features.
It updates the feature count in each data frame based on the new features provided.
}
\examples{
pca_features <- perform_pca_feature_selection(DT[train_ind, ])
pca_feature_counts <- initialize_feature_counts(pca_features)
updated_counts <- update_features_df_list(pca_feature_counts, pca_features)
}
