% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{xgboost_feature_selection}
\alias{xgboost_feature_selection}
\title{XGBoost Feature Selection}
\usage{
xgboost_feature_selection(DT, CS, S, nrounds = 100, N = 10)
}
\arguments{
\item{DT}{Data frame containing the features and outcome.}

\item{CS}{Censoring status.}

\item{S}{Survival time or binary indicator for survival.}

\item{nrounds}{Number of boosting rounds.}

\item{N}{The divisor to determine the top percentage of features to select.}
}
\value{
A data frame of top selected features based on their importance scores.
}
\description{
Performs feature selection using XGBoost and selects the top 1/N features.
}
