% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean.R
\name{boolean}
\alias{boolean}
\title{Returns a boolean object}
\usage{
boolean(x, data = 0)
}
\arguments{
\item{x}{Name of object}

\item{data}{vector (defaults to 0)}
}
\description{
\code{linear} returns an boolean object.
}
