% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract-convert.R
\name{contract_convert}
\alias{contract_convert}
\title{Convert contract names to factor intervals}
\usage{
contract_convert(x)
}
\arguments{
\item{x}{A character vector of contract names.}
}
\value{
A factor vector intervals in proper notation.
}
\description{
Convert contract names to factor intervals
}
\examples{
contract_convert(c("3\%-", "12\%+", "9\% - 12\%", "3\% - 6\%", "6\% - 9\%"))
contract_convert(c("219-", "220 - 229", "240+",  "230 - 239"))
contract_convert(c("Sanders", "Biden", "Warren", "Buttigieg"))
}
