% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept.R
\name{intercept}
\alias{intercept}
\title{Returns an intercept object}
\usage{
intercept(x = 0)
}
\arguments{
\item{x}{Intercept vector (default to 0)}
}
\description{
\code{intercept} returns an intercept object.
}
\details{
This object represents the constant, or intercept,
associated with a linear-based model.
For univariate models,
the intercept is a scalar.
For multivariate models,
the intercept is a vector.
This vector is passed to \code{intercept} using the \code{x}
parameter.
}
