% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept.R
\name{intercept}
\alias{intercept}
\title{Returns an intercept object}
\usage{
intercept(x = 0)
}
\arguments{
\item{x}{Intercept vector (default to 0)}
}
\description{
\code{intercept} returns an intercept object.
}
