% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{Returns a linear object}
\usage{
linear(x, data = 0)
}
\arguments{
\item{x}{Name of object}

\item{data}{vector (defaults to 0)}
}
\description{
\code{linear} returns an linear object.
}
