% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-history.R
\name{market_history}
\alias{market_history}
\title{Market price history}
\usage{
market_history(id, timespan = c("24h", "7d", "90d"))
}
\arguments{
\item{id}{Market ID number.}

\item{timespan}{One of "24h", "7d" or "90d". Either hourly or daily.}
}
\value{
A data frame of market contract prices over time.
}
\description{
Download a data frame of contract prices over time, either hourly or daily.
This chart data does not include a market name or numeric contract IDs, which
can be obtained from the list of all \code{\link[=open_markets]{open_markets()}}.
}
\examples{
market_history(id = 6653, timespan = "7d")
}
