% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-history.R
\name{market_history}
\alias{market_history}
\title{Market price history}
\format{
A tibble with 8 variables:
\describe{
\item{time}{The hour or day of price}
\item{mid}{Market ID}
\item{contract}{Question answers}
\item{open}{Price at start of hour/day}
\item{close}{Price at end of hour/day}
\item{low}{Low price of hour/day}
\item{high}{High price of hour/day}
\item{volume}{Number of shares traded per hour/day}
}
}
\usage{
market_history(mid, hourly = FALSE)
}
\arguments{
\item{mid}{Market ID.}

\item{hourly}{24 hourly rows per contract? If not, 90 daily rows.}
}
\value{
A data frame of market contract prices over time.
}
\description{
Download a data frame of contract prices over time, either hourly or daily.
This chart data does not include a market name or numeric contract IDs, which
can be obtained from the list of all \code{\link[=open_markets]{open_markets()}}.
}
\examples{
market_history(mid = 3698)
market_history(mid = 6653, hourly = TRUE)
}
