% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-plot.R
\name{market_plot}
\alias{market_plot}
\title{Plot market types}
\usage{
market_plot(data, color = 4)
}
\arguments{
\item{data}{A data frame created by either \code{\link[=market_history]{market_history()}} or
\code{\link[=market_price]{market_price()}}. For price history plots, \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}} is used;
for current prices, \code{\link[ggplot2:geom_col]{ggplot2::geom_col()}} is used.}

\item{color}{The number of current top contracts lines highlighted if \code{data}
is daily or hourly.}
}
\value{
A ggplot object.
}
\description{
Plot market types
}
\examples{
market_plot(market_history(3633), color = 4)
market_plot(market_price(6405))
}
