% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-price.R
\name{market_price}
\alias{market_price}
\title{Current market prices}
\usage{
market_price(id)
}
\arguments{
\item{id}{The integer market ID.}
}
\value{
A data frame of market contracts prices.
}
\description{
Current market prices
}
\examples{
market_price(id = 6653)
}
