% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-markets.R
\name{open_markets}
\alias{open_markets}
\title{List all open markets}
\format{
A tibble with 8 variables:
\describe{
\item{time}{The hour or day of price}
\item{mid}{Market ID}
\item{market}{Market question}
\item{cid}{Contract ID}
\item{contract}{Question answers}
\item{last}{Most recent trading price of the contract}
\item{close}{Price at the end of the previous midnight EST}
\item{end}{The date and time which the market closes}
}
}
\usage{
open_markets(split = FALSE)
}
\arguments{
\item{split}{Split into a list of market data frames?}
}
\value{
A data frame or list of market contracts prices.
}
\description{
List all open markets
}
\examples{
open_markets(split = FALSE)
}
