% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-markets.R
\name{open_markets}
\alias{open_markets}
\title{List all open markets}
\usage{
open_markets()
}
\value{
A data frame of markets with nested contract prices.
}
\description{
List all open markets
}
\examples{
open_markets()
}
