% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictr.R
\name{predictr}
\alias{predictr}
\title{Predict a binary event based on a model or coefficents}
\usage{
predictr(data, coefficient = NULL, fit = NULL)
}
\arguments{
\item{data}{Tibble containing all coefficents / outcome of interest as specified in fit or coefficent parameters.}

\item{coefficient}{Output from the coefficent function (alternative to fit parameter)}

\item{fit}{Logistic regression fit object (alternative to coefficient parameter)}
}
\value{
Tibble with 2 appended columns: "predict_raw" (the original output prediction) and "predict_prop" (the predicted probability of the event)
}
\description{
Predict a binary event based on a model or coefficents
}
