% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_features.R
\name{select_features}
\alias{select_features}
\title{Feature Selection using Caret's Random Forest Approach}
\usage{
select_features(x, y, n_core = 8, n_cv_fold = 10, n_cv_repeat = 1,
  n_run = 20, n_maxsamp = 1000, n_size = c(2:10), n_cutoff = 0.6,
  verbose = TRUE)
}
\arguments{
\item{x}{a data frame or matrix of predictors (x).}

\item{y}{target (numeric or factor)}
}
\description{
This function determines the best combination of predictors.
}
\examples{
## T.B.A.

}

